+++ policy_sentry/writing/sid_group.py
@@ -94,11 +94,18 @@ class SidGroup:
         for sid in self.sids:
             actions = self.sids[sid]["actions"]
             if len(actions) == 0:
+                logger.debug(f"No actions for sid {sid}")
                 continue
             if minimize is not None and isinstance(minimize, int):
+                logger.debug("Minimizing statements...")
                 actions = minimize_statement_actions(
                     actions, all_actions, minchars=minimize
                 )
+            logger.debug(f"Adding statement with SID {sid}")
+            logger.debug(f"{sid} SID has the actions: {actions}")
+            logger.debug(f"{sid} SID has the resources: {self.sids[sid]['arn']}")
+
+
             statements.append(
                 {
                     "Sid": sid,
@@ -209,7 +216,9 @@ class SidGroup:
         dependent_actions = get_dependent_actions(db_session, supplied_actions)
         # List comprehension to get all dependent actions that are not in the supplied actions.
         dependent_actions = [x for x in dependent_actions if x not in supplied_actions]
-
+        logger.debug("Adding by list of actions")
+        logger.debug(f"Supplied actions: {str(supplied_actions)}")
+        logger.debug(f"Dependent actions: {str(dependent_actions)}")
         arns_matching_supplied_actions = []
 
         # arns_matching_supplied_actions is a list of dicts.
@@ -267,10 +276,16 @@ class SidGroup:
                 # self.add_action_without_resource_constraint(str.lower(dep_action))
         # Now, because add_by_arn_and_access_level() adds all actions under an access level, we have to
         # remove all actions that do not match the supplied_actions. This is done in-place.
+        logger.debug("Purging actions that do not match the requested actions and dependent actions")
+        logger.debug(f"Supplied actions: {str(supplied_actions)}")
+        logger.debug(f"Dependent actions: {str(dependent_actions)}")
         self.remove_actions_not_matching_these(supplied_actions + dependent_actions)
         for action in actions_without_resource_constraints:
+            logger.debug(f"Deliberately adding the action {action} without resource constraints")
             self.add_action_without_resource_constraint(action)
+        logger.debug("Removing actions that are in the wildcard arn (Resources = '*') as well as other statements that have resource constraints")
         self.remove_actions_duplicated_in_wildcard_arn()
+        logger.debug("Getting the rendered policy")
         rendered_policy = self.get_rendered_policy(db_session)
         return rendered_policy
 
@@ -285,6 +300,7 @@ class SidGroup:
         # try:
         if "mode" in cfg.keys():
             if cfg["mode"] == "crud":
+                logger.debug("CRUD mode selected")
                 check_crud_schema(cfg)
                 if "wildcard-only" in cfg.keys():
                     if "single-actions" in cfg["wildcard-only"]:
@@ -293,6 +309,7 @@ class SidGroup:
                                 provided_wildcard_actions = cfg["wildcard-only"][
                                     "single-actions"
                                 ]
+                                logger.debug(f"Requested wildcard-only actions: {str(provided_wildcard_actions)}")
                                 self.add_wildcard_only_actions(
                                     db_session, provided_wildcard_actions
                                 )
@@ -300,6 +317,7 @@ class SidGroup:
                         if cfg["wildcard-only"]["service-read"]:
                             if cfg["wildcard-only"]["service-read"][0] != "":
                                 service_read = cfg["wildcard-only"]["service-read"]
+                                logger.debug(f"Requested wildcard-only actions: {str(service_read)}")
                                 self.add_wildcard_only_actions_matching_services_and_access_level(
                                     db_session, service_read, "Read"
                                 )
@@ -307,6 +325,7 @@ class SidGroup:
                         if cfg["wildcard-only"]["service-write"]:
                             if cfg["wildcard-only"]["service-write"][0] != "":
                                 service_write = cfg["wildcard-only"]["service-write"]
+                                logger.debug(f"Requested wildcard-only actions: {str(service_write)}")
                                 self.add_wildcard_only_actions_matching_services_and_access_level(
                                     db_session, service_write, "Write"
                                 )
@@ -314,6 +333,7 @@ class SidGroup:
                         if cfg["wildcard-only"]["service-list"]:
                             if cfg["wildcard-only"]["service-list"][0] != "":
                                 service_list = cfg["wildcard-only"]["service-list"]
+                                logger.debug(f"Requested wildcard-only actions: {str(service_list)}")
                                 self.add_wildcard_only_actions_matching_services_and_access_level(
                                     db_session, service_list, "List"
                                 )
@@ -323,6 +343,7 @@ class SidGroup:
                                 service_tagging = cfg["wildcard-only"][
                                     "service-tagging"
                                 ]
+                                logger.debug(f"Requested wildcard-only actions: {str(service_tagging)}")
                                 self.add_wildcard_only_actions_matching_services_and_access_level(
                                     db_session, service_tagging, "Tagging"
                                 )
@@ -334,9 +355,11 @@ class SidGroup:
                                 ]
                                 != ""
                             ):
+
                                 service_permissions_management = cfg["wildcard-only"][
                                     "service-permissions-management"
                                 ]
+                                logger.debug(f"Requested wildcard-only actions: {str(service_permissions_management)}")
                                 self.add_wildcard_only_actions_matching_services_and_access_level(
                                     db_session,
                                     service_permissions_management,
@@ -344,16 +367,19 @@ class SidGroup:
                                 )
                 if "read" in cfg.keys():
                     if cfg["read"] is not None and cfg["read"][0] != "":
+                        logger.debug(f"Requested access to arns: {str(cfg['read'])}")
                         self.add_by_arn_and_access_level(
                             db_session, cfg["read"], "Read"
                         )
                 if "write" in cfg.keys():
                     if cfg["write"] is not None and cfg["write"][0] != "":
+                        logger.debug(f"Requested access to arns: {str(cfg['write'])}")
                         self.add_by_arn_and_access_level(
                             db_session, cfg["write"], "Write"
                         )
                 if "list" in cfg.keys():
                     if cfg["list"] is not None and cfg["list"][0] != "":
+                        logger.debug(f"Requested access to arns: {str(cfg['list'])}")
                         self.add_by_arn_and_access_level(
                             db_session, cfg["list"], "List"
                         )
@@ -362,6 +388,7 @@ class SidGroup:
                         cfg["permissions-management"] is not None
                         and cfg["permissions-management"][0] != ""
                     ):
+                        logger.debug(f"Requested access to arns: {str(cfg['permissions-management'])}")
                         self.add_by_arn_and_access_level(
                             db_session,
                             cfg["permissions-management"],
@@ -369,6 +396,7 @@ class SidGroup:
                         )
                 if "tagging" in cfg.keys():
                     if cfg["tagging"] is not None and cfg["tagging"][0] != "":
+                        logger.debug(f"Requested access to arns: {str(cfg['tagging'])}")
                         self.add_by_arn_and_access_level(
                             db_session, cfg["tagging"], "Tagging"
                         )
@@ -394,11 +422,9 @@ class SidGroup:
                 db_session, provided_wildcard_actions
             )
             if len(verified_wildcard_actions) > 0:
+                logger.debug("Attempting to add the following actions to the policy: %s", verified_wildcard_actions)
                 self.add_by_list_of_actions(db_session, verified_wildcard_actions)
-                logger.debug(
-                    "Added the following wildcard-only actions to the policy: %s",
-                    verified_wildcard_actions,
-                )
+                logger.debug("Added the following wildcard-only actions to the policy: %s", verified_wildcard_actions)
 
     def add_wildcard_only_actions_matching_services_and_access_level(
         self, db_session, services, access_level