+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ## Future release (Unreleased)
 * Add conditions support
 
+## 0.7.1.3 (Unreleased)
+* Better logging for the write-policy command
+* Fixes #142 - the issue with validating matching resource ARN types
+
 ## 0.7.1.2 (2020-03-08)
 * Minor fix for `kmsKey` ARN type under `imagebuilder` service. Fixes #114
 * Added PyInvoke command for updating the aws documentation.